requirements.txt
biopython>=1.84
contractions==0.1.73
nltk>=3.9.1
num2words>=0.5.14
numpy>=1.24.3
pandas>=2.0.3
pyenchant==3.2.2
pylangacq==0.19.1
python-Levenshtein>=0.27.1
PyYAML>=6.0.2
selenium>=4.27.1
tqdm>=4.67.1
openpyxl>=3.1.5
streamlit>=1.45.1
scipy>=1.16.0